requirements.txt
asgiref==3.3.4
certifi==2023.5.7
channels-redis==3.4.0
channels==3.0.4
charset-normalizer==2.0.7
daphne==3.0.2
django-cities-light==3.9.0
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-select2==7.10.1
django-simple-sso==1.1.0
django-webpack-loader==1.8.1
django-background-tasks==1.2.5
Django==3.2.9
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
djlint==1.29.0
Faker==9.8.2
gunicorn==19.9.0
idna==3.3
itsdangerous==0.24
psycopg2==2.9.1
pytz==2021.3
requests==2.31.0
sqlparse==0.4.4
typing-extensions==3.10.0.2
urllib3==1.26.7
webservices==0.7
# Our custom package for the S3 Provider must be the last line of this file to avoid github actions issues
git+https://github.com/Insight-Scope/aws-s3-provider@main#=aws-s3-provider
